Peace has been restored to the universe.
I messed up alot with my pulls recently; I accindetally pulled on one machine with changes on another (without the knowledge that I could discard changes on the other machine so then I didn't have to go through the painful and often code-breaking process of 'merging' the code). This caused a big panic as code was mixed and matched, so I had to go back to the last 'good' patch (it was one before I tried to merge with that good patch, and then managed to do that poorly then immediatley push before realising that it was a mess). Basically, big jumble up, re-did the fixes that I did, and here we are. Actual notes as to where it's at:

Weatherapp div is currently 'sticky' to the top of the browser, so if you have a long webpage, it stays up there as you scroll, although I'm thinking we ditch this anyways as we can just have everything on one page, as long as we can get it to properly fully fill the display area without any weird issues that cause it to be too long or too short.

The colour schemeing needs some finalising too on it, as the red and blue is nice but I don't think fits.

There's a line to mark the end of the first div and start of the second div (mainly just for reference). The colour shifts to blue because I was thinking of having that as the background of the dynamic weather graphic (as it would be the 'blue sky', although with some techinical CSS that i can learn in the future we can make it shift colour to be darker for storms or cooler weather, or maybe even based on time. but that's later down the line.

that's mainly all i've done, one or two small tweaks and i fucked up a shit ton before fixing it. fun!