wholesome dupe checker in search history
added a neat little check that cross-matches the current search with each entry in the history box.
if it finds a matches, it will break the counter early.
checks at end if it has run 9 times without breaking due to dupe, and if so will add that to the recent search.

also stopped the 100 console logs of api data